防重复命名调整、防重复增加时间段限制

+++ eva-server/src/main/java/com/eva/config/shiro/ShiroConfig.java
@@ -25,7 +25,7 @@ import java.util.Map;
 public class ShiroConfig {
 
     @Value("${cache.session.expire}")
-    private long sessionExpireTime;
+    private int sessionExpireTime;
 
     @Autowired
     private ShiroCredentialsMatcher shiroCredentialsMatcher;
